Problem: no negative test for state check
